Refactor secrets command
Move all secrets command code into cmd package.
